requirements.txt
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
python-env==1.0.0
requests==2.31.0
soupsieve==2.5
urllib3==2.0.6
Werkzeug==3.0.0
